Move exql to internal.
