+++ internal/sqladapter/collection.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 
 	"upper.io/db.v2"
-	"upper.io/db.v2/lib/sqlbuilder/exql"
+	"upper.io/db.v2/internal/sqladapter/exql"
 )
 
 // Collection represents a SQL table.
