+++ internal/sqladapter/database.go
@@ -9,8 +9,8 @@ import (
 	"upper.io/db.v2"
 	"upper.io/db.v2/internal/cache"
 	"upper.io/db.v2/internal/logger"
+	"upper.io/db.v2/internal/sqladapter/exql"
 	"upper.io/db.v2/lib/sqlbuilder"
-	"upper.io/db.v2/lib/sqlbuilder/exql"
 )
 
 // HasCleanUp
