+++ internal/sqladapter/exql/column.go
