+++ internal/sqladapter/exql/column_test.go
