+++ internal/sqladapter/exql/columns.go
