+++ internal/sqladapter/exql/columns_test.go
