+++ internal/sqladapter/exql/group_by_test.go
