+++ internal/sqladapter/exql/interfaces.go
