+++ internal/sqladapter/exql/table.go
