+++ internal/sqladapter/exql/utilities.go
