+++ internal/sqladapter/exql/value_test.go
