+++ internal/sqladapter/exql/where.go
