+++ internal/sqladapter/exql/where_test.go
