+++ lib/sqlbuilder/convert.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"upper.io/db.v2"
-	"upper.io/db.v2/lib/sqlbuilder/exql"
+	"upper.io/db.v2/internal/sqladapter/exql"
 )
 
 var (
