+++ lib/sqlbuilder/select.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"upper.io/db.v2"
-	"upper.io/db.v2/lib/sqlbuilder/exql"
+	"upper.io/db.v2/internal/sqladapter/exql"
 )
 
 type selectMode uint8
