+++ mysql/template.go
@@ -23,7 +23,7 @@ package mysql
 
 import (
 	"upper.io/db.v2/internal/cache"
-	"upper.io/db.v2/lib/sqlbuilder/exql"
+	"upper.io/db.v2/internal/sqladapter/exql"
 )
 
 const (
